make NewFromInt to accept int as type

+++ decimal.go
@@ -93,9 +93,9 @@ func New(value int64, exp int32) Decimal {
 //
 //     NewFromInt(123).String() // output: "123"
 //     NewFromInt(-10).String() // output: "-10"
-func NewFromInt(value int64) Decimal {
+func NewFromInt(value int) Decimal {
 	return Decimal{
-		value: big.NewInt(value),
+		value: big.NewInt(int64(value)), 
 		exp:   0,
 	}
 }
